Move postRequest function to databaseAPI.
